cai-nav: Generalized the use of the PathCorridorData and CornerData classes.
Renamed CrowdCornerData to CornerData so it is suitable for use with both the CrowdManager and PathCorridor classes.
Refactored the PathCorridor related classes to use CornerData instead of individual array buffers.
Refactored the PathCorridorData and CornerData classes so they are no longer fixed size buffers. (Though a fixed size is still required for interop method calls.)