CAINav: Refactored from using abstract classes for Unity extendability to interfaces.
The best design for the Unity extendability features is to use an abstract class model.  But a critical, long term defect in Unity prevents that type of model.  So extensions are now based on interfaces.  Specifically, the CAINavmeshData, InputBuildProcessor, and SceneQuery abstract classes have been converted to interfaces.

The Unity defect: A concrete class in a compiled library that extends an abstract class in another library will not be recognized by various Unity features if the abstract class extends either Component or ScriptableObject.  Defect example (pre-change): The AreaMarkerCompiler (in cai-nmbuild-extras-u3d) was invisible because it extended InputBuildProcessor (in cai-nmbuild-u3d).  This defect is only triggered if both elements are in compiled libraries, rather than scripts.